pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[dependency-groups]
dev = [
  "ab-test-fixtures>=0.1.5,<0.2.0",
  "debugpy>=1.8.14,<2",
  "isort>=6.0.1,<7",
  "pre-commit>=4.2.0,<5",
  "pytest-asyncio>=1.0.0,<2",
  "pytest-cov>=6.2.1,<7",
  "pytest>=8.4.1,<9",
  "ruff>=0.12.3,<0.13",
  "tox>=4.27.0,<5"
]

[project]
authors = [{email = "53892067+mattcoulter7@users.noreply.github.com", name = "Matthew Coulter"}]
dependencies = [
  "ab-alembic-auto-migrate>=0.2.0,<0.3.0",
  "ab-database[sqlite-async,postgres-sync,postgres-async]>=0.2.2,<0.3.0",
  "ab-logging>=0.1.2,<0.2.0",
  "ab-identity-context[fastapi]>=0.2.0,<0.3.0",
  "ab-client-token-validator>=1.0.0,<2",
  "ab-client-user>=1.0.0,<2",
  "ab-client-auth-client>=1.0.0,<2",
  "ab-client-token-issuer>=1.0.0,<2",
  "ab-client-token-issuer-store>=1.0.0,<2",
  "ab-client-token-store>=1.0.0,<2",
  "fastapi>=0.115.14,<0.116",
  "pydantic[email]>=2.11.7,<3",
  "uvicorn>=0.35.0,<0.36",
]
description = "Auth Broker BFF Service."
name = "ab-bff-service"
readme = "README.md"
requires-python = ">=3.12, <4"
version = "0.3.0"

[tool.hatch.build.targets.sdist]
include = ["src/ab_service"]

[tool.hatch.build.targets.wheel]
include = ["src/ab_service"]

[tool.hatch.build.targets.wheel.sources]
"src/ab_service" = "ab_service"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
asyncio_mode = "auto"
markers = [
  "integration: Assigns the test to the integration test suite.",
  "regression: Assigns the test to the regression test suite"
]

[tool.ruff]
line-length = 120
src = ["src"]
target-version = "py312"

[tool.ruff.lint]
exclude = [".git", ".venv", "__pycache__", "proto"]
fixable = ["ALL"]
ignore = [
  "D104" # missing docstring in public package
]
select = [
  "ARG001", # unused arguments
  "B", # flake8-bugbear
  "C4", # flake8-comprehensions
  "D", # pydocstyle (docstring checks)
  "E", # pycodestyle errors
  "F", # pyflakes
  "I", # isort
  "UP", # pyupgrade
  "W" # pycodestyle warnings
]
